Fix compilation on 32-bit systems
storesession.h was declaring fields as std::atomic<uint64_t>, a type
not avaiable on 32-bit systems.  However, the values stored in these
fields were only used as input to QProgressDialog, which takes int,
not uint64_t.  So the fields could just as well be std::atomic<int>.

Also, added code to scale the progress values down if they would not
fit in an int.  (This would have been needed even if the fields were
to remain as uint64_t.)

+++ pv/dialogs/storeprogress.cpp
@@ -66,7 +66,7 @@ void StoreProgress::closeEvent(QCloseEvent*)
 
 void StoreProgress::on_progress_updated()
 {
-	const std::pair<uint64_t, uint64_t> p = _session.progress();
+	const std::pair<int, int> p = _session.progress();
 	assert(p.first <= p.second);
 
 	if (p.second) {
